[AIRFLOW-5190] Move GCP Compute to core (#5797)
