+++ tests/gcp/operators/test_compute_system_helper.py
